Moved certain traits and classes to top-level in respective packages.
