+++ src/main/scala/org/economicsl/auctions/multiunit/package.scala
@@ -15,5 +15,6 @@ limitations under the License.
 */
 package org.economicsl.auctions
 
+
 /** Documentation for Multi-Unit Auction goes here! */
 package object multiunit
